🩹 Fix: 유효성 검사 오류 해결
